Unify admin ID handling, add worker healthcheck, and use Decimal for monetary fields  - security: add get_admin_ids(); rely on is_admin_uid for admin checks - start handler: remove manual TELEGRAM_ADMIN_IDS parsing; fix channel gate indentation - compose: add basic healthcheck to worker service - models: change monetary fields to Decimal for Numeric(12,2) compatibility
Changes:
1) Admin IDs
   - app/services/security.py: new get_admin_ids() helper.
   - start.py: switched to is_admin_uid; removed duplicate ENV parsing.
   - Fixed indentation issues introduced during previous edits in channel gate block.

2) Worker healthcheck
   - docker-compose.yml: added a lightweight healthcheck that imports the
     scheduler module and prints 'ok', with sensible interval/timeout/retries.

3) Monetary typing
   - app/db/models.py: monetary fields now use Decimal instead of float to align
     with Numeric(12,2) and prevent precision issues.

Impact:
- More maintainable and consistent admin authorization code.
- Better operational visibility of the worker service in Compose.
- Improved financial correctness and type consistency across the codebase.

Notes:
- No DB migrations required (schema unchanged).
- If mypy/linters are used, this will reduce type warnings for monetary fields.
- Consider gradually refactoring other modules that manually parse admin IDs to use the security module as well (plans/orders/wallet if any).

+++ app/db/models.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from datetime import datetime
 from typing import Optional
+from decimal import Decimal
 
 from sqlalchemy import BigInteger, Boolean, DateTime, Enum, ForeignKey, Integer, Numeric, String, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
@@ -26,7 +27,7 @@ class User(Base):
     last_notified_usage_threshold: Mapped[Optional[float]] = mapped_column(Numeric(5, 4), nullable=True)
     last_notified_expiry_day: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
 
-    balance: Mapped[float] = mapped_column(Numeric(12, 2), default=0)
+    balance: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
 
     created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)
     updated_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
@@ -38,7 +39,7 @@ class Plan(Base):
     id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
     template_id: Mapped[int] = mapped_column(Integer, unique=True, index=True)
     title: Mapped[str] = mapped_column(String(191))
-    price: Mapped[float] = mapped_column(Numeric(12, 2))
+    price: Mapped[Decimal] = mapped_column(Numeric(12, 2))
     currency: Mapped[str] = mapped_column(String(8), default="IRR")
     duration_days: Mapped[int] = mapped_column(Integer, default=30)
     data_limit_bytes: Mapped[int] = mapped_column(BigInteger)
@@ -57,13 +58,13 @@ class Order(Base):
     # Snapshot of plan details at purchase time (decoupled from Plan lifecycle)
     plan_template_id: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
     plan_title: Mapped[Optional[str]] = mapped_column(String(191), nullable=True)
-    plan_price: Mapped[Optional[float]] = mapped_column(Numeric(12, 2), nullable=True)
+    plan_price: Mapped[Optional[Decimal]] = mapped_column(Numeric(12, 2), nullable=True)
     plan_currency: Mapped[Optional[str]] = mapped_column(String(8), nullable=True)
     plan_duration_days: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
     plan_data_limit_bytes: Mapped[Optional[int]] = mapped_column(BigInteger, nullable=True)
 
     status: Mapped[str] = mapped_column(String(32), index=True, default="pending")
-    amount: Mapped[float] = mapped_column(Numeric(12, 2))
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 2))
     currency: Mapped[str] = mapped_column(String(8), default="IRR")
     provider: Mapped[str] = mapped_column(String(32), default="manual_transfer")
     provider_ref: Mapped[Optional[str]] = mapped_column(String(191), nullable=True)
